+++ src/main/java/cc/altius/FASP/model/rowMapper/SimplifiedSupplyPlanResultSetExtractor.java
@@ -28,46 +28,46 @@ public class SimplifiedSupplyPlanResultSetExtractor implements ResultSetExtracto
             sp.setConversionFactor(rs.getDouble("CONVERSION_FACTOR"));
             int idx = spList.indexOf(sp);
             if (idx == -1) {
-                sp.setOpeningBalance(rs.getLong("OPENING_BALANCE"));
-                sp.setOpeningBalanceWps(rs.getLong("OPENING_BALANCE_WPS"));
+                sp.setOpeningBalance(rs.getDouble("OPENING_BALANCE"));
+                sp.setOpeningBalanceWps(rs.getDouble("OPENING_BALANCE_WPS"));
                 sp.setActualFlag(rs.getBoolean("ACTUAL_FLAG"));
                 if (rs.wasNull()) {
                     sp.setActualFlag(null);
                 }
-                sp.setConsumptionQty(rs.getLong("CONSUMPTION_QTY"));
+                sp.setConsumptionQty(rs.getDouble("CONSUMPTION_QTY"));
                 if (rs.wasNull()) {
                     sp.setConsumptionQty(null);
                 }
-                sp.setAdjustmentQty(rs.getLong("ADJUSTMENT_MULTIPLIED_QTY"));
+                sp.setAdjustmentQty(rs.getDouble("ADJUSTMENT_MULTIPLIED_QTY"));
                 if (rs.wasNull()) {
                     sp.setAdjustmentQty(null);
                 }
-                sp.setStockQty(rs.getLong("STOCK_MULTIPLIED_QTY"));
+                sp.setStockQty(rs.getDouble("STOCK_MULTIPLIED_QTY"));
                 if (rs.wasNull()) {
                     sp.setStockQty(null);
                 }
                 sp.setRegionCount(rs.getInt("REGION_COUNT"));
                 sp.setRegionCountForStock(rs.getInt("REGION_COUNT_FOR_STOCK"));
-                sp.setPlannedShipmentsTotalData(rs.getLong("MANUAL_PLANNED_SHIPMENT_QTY"));
-                sp.setSubmittedShipmentsTotalData(rs.getLong("MANUAL_SUBMITTED_SHIPMENT_QTY"));
-                sp.setApprovedShipmentsTotalData(rs.getLong("MANUAL_APPROVED_SHIPMENT_QTY"));
-                sp.setShippedShipmentsTotalData(rs.getLong("MANUAL_SHIPPED_SHIPMENT_QTY"));
-                sp.setReceivedShipmentsTotalData(rs.getLong("MANUAL_RECEIVED_SHIPMENT_QTY"));
-                sp.setOnholdShipmentsTotalData(rs.getLong("MANUAL_ONHOLD_SHIPMENT_QTY"));
-                sp.setPlannedErpShipmentsTotalData(rs.getLong("ERP_PLANNED_SHIPMENT_QTY"));
-                sp.setSubmittedErpShipmentsTotalData(rs.getLong("ERP_SUBMITTED_SHIPMENT_QTY"));
-                sp.setApprovedErpShipmentsTotalData(rs.getLong("ERP_APPROVED_SHIPMENT_QTY"));
-                sp.setShippedErpShipmentsTotalData(rs.getLong("ERP_SHIPPED_SHIPMENT_QTY"));
-                sp.setReceivedErpShipmentsTotalData(rs.getLong("ERP_RECEIVED_SHIPMENT_QTY"));
-                sp.setOnholdErpShipmentsTotalData(rs.getLong("ERP_ONHOLD_SHIPMENT_QTY"));
-                sp.setExpiredStock(rs.getLong("EXPIRED_STOCK"));
-                sp.setExpiredStockWps(rs.getLong("EXPIRED_STOCK_WPS"));
-                sp.setClosingBalance(rs.getLong("CLOSING_BALANCE"));
-                sp.setClosingBalanceWps(rs.getLong("CLOSING_BALANCE_WPS"));
-                sp.setUnmetDemand(rs.getLong("UNMET_DEMAND"));
-                sp.setUnmetDemandWps(rs.getLong("UNMET_DEMAND_WPS"));
-                sp.setNationalAdjustment(rs.getLong("NATIONAL_ADJUSTMENT"));
-                sp.setNationalAdjustmentWps(rs.getLong("NATIONAL_ADJUSTMENT_WPS"));
+                sp.setPlannedShipmentsTotalData(rs.getDouble("MANUAL_PLANNED_SHIPMENT_QTY"));
+                sp.setSubmittedShipmentsTotalData(rs.getDouble("MANUAL_SUBMITTED_SHIPMENT_QTY"));
+                sp.setApprovedShipmentsTotalData(rs.getDouble("MANUAL_APPROVED_SHIPMENT_QTY"));
+                sp.setShippedShipmentsTotalData(rs.getDouble("MANUAL_SHIPPED_SHIPMENT_QTY"));
+                sp.setReceivedShipmentsTotalData(rs.getDouble("MANUAL_RECEIVED_SHIPMENT_QTY"));
+                sp.setOnholdShipmentsTotalData(rs.getDouble("MANUAL_ONHOLD_SHIPMENT_QTY"));
+                sp.setPlannedErpShipmentsTotalData(rs.getDouble("ERP_PLANNED_SHIPMENT_QTY"));
+                sp.setSubmittedErpShipmentsTotalData(rs.getDouble("ERP_SUBMITTED_SHIPMENT_QTY"));
+                sp.setApprovedErpShipmentsTotalData(rs.getDouble("ERP_APPROVED_SHIPMENT_QTY"));
+                sp.setShippedErpShipmentsTotalData(rs.getDouble("ERP_SHIPPED_SHIPMENT_QTY"));
+                sp.setReceivedErpShipmentsTotalData(rs.getDouble("ERP_RECEIVED_SHIPMENT_QTY"));
+                sp.setOnholdErpShipmentsTotalData(rs.getDouble("ERP_ONHOLD_SHIPMENT_QTY"));
+                sp.setExpiredStock(rs.getDouble("EXPIRED_STOCK"));
+                sp.setExpiredStockWps(rs.getDouble("EXPIRED_STOCK_WPS"));
+                sp.setClosingBalance(rs.getDouble("CLOSING_BALANCE"));
+                sp.setClosingBalanceWps(rs.getDouble("CLOSING_BALANCE_WPS"));
+                sp.setUnmetDemand(rs.getDouble("UNMET_DEMAND"));
+                sp.setUnmetDemandWps(rs.getDouble("UNMET_DEMAND_WPS"));
+                sp.setNationalAdjustment(rs.getDouble("NATIONAL_ADJUSTMENT"));
+                sp.setNationalAdjustmentWps(rs.getDouble("NATIONAL_ADJUSTMENT_WPS"));
                 sp.setAmc(rs.getDouble("AMC"));
                 if (rs.wasNull()) {
                     sp.setAmc(null);
@@ -96,32 +96,32 @@ public class SimplifiedSupplyPlanResultSetExtractor implements ResultSetExtracto
                         rs.getString("BATCH_NO"),
                         rs.getDate("EXPIRY_DATE"),
                         rs.getBoolean("AUTO_GENERATED"),
-                        rs.getLong("BATCH_CLOSING_BALANCE"),
-                        rs.getLong("BATCH_CLOSING_BALANCE_WPS"),
-                        rs.getLong("BATCH_EXPIRED_STOCK"),
-                        rs.getLong("BATCH_EXPIRED_STOCK_WPS"),
+                        rs.getDouble("BATCH_CLOSING_BALANCE"),
+                        rs.getDouble("BATCH_CLOSING_BALANCE_WPS"),
+                        rs.getDouble("BATCH_EXPIRED_STOCK"),
+                        rs.getDouble("BATCH_EXPIRED_STOCK_WPS"),
                         rs.getDate("BATCH_CREATED_DATE"),
-                        rs.getLong("BATCH_SHIPMENT_QTY"),
-                        rs.getLong("BATCH_SHIPMENT_QTY_WPS"));
-                sb.setOpeningBalance(rs.getLong("BATCH_OPENING_BALANCE"));
-                sb.setOpeningBalanceWps(rs.getLong("BATCH_OPENING_BALANCE_WPS"));
-                sb.setConsumptionQty(rs.getLong("BATCH_CONSUMPTION_QTY"));
+                        rs.getDouble("BATCH_SHIPMENT_QTY"),
+                        rs.getDouble("BATCH_SHIPMENT_QTY_WPS"));
+                sb.setOpeningBalance(rs.getDouble("BATCH_OPENING_BALANCE"));
+                sb.setOpeningBalanceWps(rs.getDouble("BATCH_OPENING_BALANCE_WPS"));
+                sb.setConsumptionQty(rs.getDouble("BATCH_CONSUMPTION_QTY"));
                 if (rs.wasNull()) {
                     sb.setConsumptionQty(null);
                 }
-                sb.setStockQty(rs.getLong("BATCH_STOCK_MULTIPLIED_QTY"));
+                sb.setStockQty(rs.getDouble("BATCH_STOCK_MULTIPLIED_QTY"));
                 if (rs.wasNull()) {
                     sb.setStockQty(null);
                 }
-                sb.setAdjustmentQty(rs.getLong("BATCH_ADJUSTMENT_MULTIPLIED_QTY"));
+                sb.setAdjustmentQty(rs.getDouble("BATCH_ADJUSTMENT_MULTIPLIED_QTY"));
                 if (rs.wasNull()) {
                     sb.setAdjustmentQty(null);
                 }
-                sb.setUnallocatedQty(rs.getLong("BATCH_CALCULATED_CONSUMPTION_QTY"));
+                sb.setUnallocatedQty(rs.getDouble("BATCH_CALCULATED_CONSUMPTION_QTY"));
                 if (rs.wasNull()) {
                     sb.setUnallocatedQty(null);
                 }
-                sb.setUnallocatedQtyWps(rs.getLong("BATCH_CALCULATED_CONSUMPTION_QTY_WPS"));
+                sb.setUnallocatedQtyWps(rs.getDouble("BATCH_CALCULATED_CONSUMPTION_QTY_WPS"));
                 if (rs.wasNull()) {
                     sb.setUnallocatedQtyWps(null);
                 }
